Added pickUp Object Reset, Fixed Room2Puzzle2
Pickup will now track the pickup location and move the object back to that location if it senses the object has fallen through the walls. Will NOT fix an item that is dropped into another room.
Puzzle2 moved to a central script to track progress to enable solve-condition.
Puzzle2 lighting fixed.
Room2 geometry adjusted